+++ src/context.ts
@@ -1,5 +1,4 @@
-import type { Options } from './types'
-
+import type { Options } from '~/types'
 import {
   addTrailingS3Sep,
 } from './helpers'