+++ src/uploader.ts
@@ -1,23 +1,23 @@
-import fs from 'node:fs'
-import { resolve } from 'node:path'
-import type { ResolvedConfig } from 'vite'
 import type { PutObjectCommandOutput } from '@aws-sdk/client-s3'
+import type { ResolvedConfig } from 'vite'
+import type { File, Options } from '~/types'
+import fs from 'node:fs'
 
+import { resolve } from 'node:path'
 import { S3 } from '@aws-sdk/client-s3'
+
+import { mapValues } from 'es-toolkit'
 import { lookup } from 'mime-types'
 
-import mapValues from 'lodash/mapValues'
-import type { File, Options } from './types'
-
-import { logResult } from './log'
-
 import {
-  DEFAULT_UPLOAD_OPTIONS,
-  UPLOAD_IGNORES,
   addSeperatorToPath,
+  DEFAULT_UPLOAD_OPTIONS,
   getDirectoryFilesRecursive,
   testRule,
-} from './helpers'
+  UPLOAD_IGNORES,
+} from '~/helpers'
+
+import { logResult } from '~/log'
 
 export default class Uploader {
   options: Options